// src/app/shared/models/model.ts
export interface pageSelection {
  skip: number;
  limit: number;
}

export interface apiResultFormat {
  data: [];
  totalData: number;
}
// ...
export interface instructorStudentList {
  sNo?:number;
  studentId: string;
  img: string;
  studentName: string;
  enrollDate: string;
  progress: string;
  courses: string;
  width: string;
}